feat(client): link sidebar logo to report
