+++ src/components/client/sidebar/ClientSidebar.tsx
@@ -1,13 +1,15 @@
 'use client';
 
-import { useState } from 'react';
-import { usePathname } from 'next/navigation';
+import { useMemo, useState } from 'react';
+import { useParams, usePathname } from 'next/navigation';
 import { ReportSelector } from '@/components/client/sidebar/ReportSelector';
 import { PdfDownloadButton } from '@/components/client/sidebar/PdfDownloadButton';
 import { ServiceUpgradeButton } from '@/components/client/sidebar/ServiceUpgradeButton';
 import { SidebarLogo } from '@/components/client/sidebar/SidebarLogo';
 import { SidebarMainNav } from '@/components/client/sidebar/SidebarMainNav';
 import { SidebarUserInfo } from '@/components/client/sidebar/SidebarUserInfo';
+import { useReports } from '@/hooks/workspace/useWorkspaceQuery';
+import { useLastReportStore } from '@/store/lastReport';
 import type { AuthUser } from '@/types/auth';
 
 interface ClientSidebarProps {
@@ -16,15 +18,33 @@ interface ClientSidebarProps {
 
 export function ClientSidebar({ user = null }: ClientSidebarProps) {
   const [isOpen, setIsOpen] = useState(true);
+  const params = useParams();
   const pathname = usePathname() ?? '';
+  const workspaceId = (params?.workspaceId as string | undefined) ?? '';
+  const { data: reports } = useReports(workspaceId);
+  const lastReportId = useLastReportStore((s) => s.lastReportByWorkspace[workspaceId]);
   const onReportPath = pathname.startsWith('/report/');
+
+  // 보고서 이동 URL — 현재 URL reportId 유지 → 마지막으로 본 보고서(존재 검증) → 최신 보고서
+  const reportHref = useMemo(() => {
+    if (!workspaceId) return '';
+    const currentReportId = params?.reportId as string | undefined;
+    const savedId =
+      lastReportId && reports?.some((r) => r.id === lastReportId) ? lastReportId : undefined;
+    const targetId = currentReportId ?? savedId ?? reports?.[0]?.id;
+    return targetId ? `/report/${workspaceId}/${targetId}` : '';
+  }, [workspaceId, params?.reportId, reports, lastReportId]);
 
   return (
     <aside
       className={`${isOpen ? 'w-60 px-2' : 'w-14'} border-r border-border-light bg-bg-white flex flex-col shrink-0 transition-all duration-300`}
     >
-      <SidebarLogo isOpen={isOpen} onToggle={() => setIsOpen((v) => !v)} />
-      <SidebarMainNav isOpen={isOpen} />
+      <SidebarLogo
+        isOpen={isOpen}
+        reportHref={reportHref}
+        onToggle={() => setIsOpen((v) => !v)}
+      />
+      <SidebarMainNav isOpen={isOpen} workspaceId={workspaceId} reportHref={reportHref} />
       {isOpen && (
         <div className="flex flex-col gap-2 w-full items-center mb-6">
           {/* PDF 다운로드 / 지난 보고서 — 보고서 메뉴에서만 노출 */}
